+++ src/zenml/models/__init__.py
@@ -389,7 +389,7 @@ from zenml.models.v2.misc.stack_deployment import (
     StackDeploymentConfig,
     StackDeploymentInfo,
 )
-from zenml.models.v2.misc.full_stack import (
+from zenml.models.v2.misc.info_models import (
     ComponentInfo,
     ServiceConnectorInfo,
     ServiceConnectorResourcesInfo,
